backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.2
python-multipart==0.0.20
seaborn==0.13.2
python-dotenv==1.1.1
google-generativeai==0.8.5
gTTS==2.5.4
